open class MenuItemList(name: String) : NamedList<MenuItemList>(name) Пока всё работает)
