+++ Readme.md
@@ -1,3 +1,5 @@
-# Вариант № 4 :
-Наследуемся напрямую от ArrayList
+# Вариант № 2 :
+open class MenuItemList(name: String) : NamedList<MenuItemList>(name)
+Пока всё работает)
+
