+++ src/main/kotlin/NamedList.kt
@@ -1,0 +1,5 @@
+open class NamedList<T> (val name: String) : ArrayList<T>() {
+    override fun toString(): String {
+        return name + super.toString()
+    }
+}
